update robot script
